==> aeroespacial/notebooks/1_explora_dados.ipynb <==
# Explorando os dados

# 1 - Lendo os dados

Para a modelagem, será utilizado apenas os dados referente ao target $r$, que é a razão tensorial-escalar que armazena a informação sobre ondas gravitacionais primordiais.

import pandas as pd

df = pd.read_csv("../data/02_intermediate/training_b_df.csv")

df

# 2 - Remoção de duplicados

# Verifica quais linhas são duplicatas
duplicadas = df.duplicated()
print(duplicadas)

total_duplicadas = df.duplicated().sum()
print(f"Total de linhas duplicadas: {total_duplicadas}")

### OBS: Não há dados duplicados neste dataset

# 3 - Verifica existencia de dados nulos

contagem_nulos = df.isnull().sum()
print("Contagem de nulos por coluna:")
print(contagem_nulos)

total_nulos_df = contagem_nulos.sum()

print(f"O número TOTAL de nulos no DataFrame é: {total_nulos_df}")

### OBS: Não há dados nulos no dataset.

# 4 - Checando tipos das colunas

print("Informações detalhadas do DataFrame:")
df.info()

### OBS: Todos os dados são númericos do tipo float64


# 5 - Verificando se os dados estão normalizados

maximo_global = df.max(numeric_only=True).max()

print(f"O valor máximo de todo o DataFrame é: {maximo_global}")

minimo_global = df.max(numeric_only=True).min()

print(f"O valor mínimo de todo o DataFrame é: {minimo_global}")

maximo_target = df["r"].max()

print(f"O valor máximo do target é: {maximo_target}")

minimo_target = df["r"].min()

print(f"O valor mínimo do target é: {minimo_target}")

### OBS:

- Os dados númericos todos se encontram entre 0 e 1, mas não estão normalizados.

- A normalização será feita após dividir-se o dataset entre X_features e y_target no próximo notebook.


# Verificando dados de teste

# 1 - Lendo os dados

Para a modelagem, será utilizado apenas os dados referente ao target $r$, que é a razão tensorial-escalar que armazena a informação sobre ondas gravitacionais primordiais.

import pandas as pd

df = pd.read_csv("../data/02_intermediate/teste_b_df.csv")

df

# 2 - Remoção de duplicados

# Verifica quais linhas são duplicatas
duplicadas = df.duplicated()
print(duplicadas)

total_duplicadas = df.duplicated().sum()
print(f"Total de linhas duplicadas: {total_duplicadas}")

### OBS: Não há dados duplicados neste dataset

# 3 - Verifica existencia de dados nulos

contagem_nulos = df.isnull().sum()
print("Contagem de nulos por coluna:")
print(contagem_nulos)

total_nulos_df = contagem_nulos.sum()

print(f"O número TOTAL de nulos no DataFrame é: {total_nulos_df}")

### OBS: Não há dados nulos no dataset.

# 4 - Checando tipos das colunas

print("Informações detalhadas do DataFrame:")
df.info()

### OBS: Todos os dados são númericos do tipo float64


# 5 - Verificando se os dados estão normalizados

maximo_global = df.max(numeric_only=True).max()

print(f"O valor máximo de todo o DataFrame é: {maximo_global}")

minimo_global = df.max(numeric_only=True).min()

print(f"O valor mínimo de todo o DataFrame é: {minimo_global}")

maximo_target = df["r"].max()

print(f"O valor máximo do target é: {maximo_target}")

minimo_target = df["r"].min()

print(f"O valor mínimo do target é: {minimo_target}")

### OBS:

- Os dados númericos todos se encontram entre 0 e 1, mas não estão normalizados.

- A normalização será feita nos próximos notebooks.
